requirements.txt
requests==2.22.0
Pillow==7.0.0
